dags/kpmg-pipeline/text_extractor/requirements.txt
algoliasearch==2.6.2
azure-core==1.26.1
azure-storage-blob==12.14.1
camelot-py==0.9.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
et-xmlfile==1.1.0
idna==3.4
isodate==0.6.1
langdetect==1.0.9
msrest==0.7.1
numpy==1.24.0
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.2
pdfminer.six==20221105
pycparser==2.21
PyPDF2==2.12.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
